+++ app/src/main/java/com/forrestguice/suntimeswidget/layouts/MoonLayout_1x1_1.java
@@ -101,7 +101,7 @@ public class MoonLayout_1x1_1 extends MoonLayout
 
     @SuppressWarnings("EmptyMethod")
     @Override
-    public void prepareForUpdate(SuntimesMoonData data)
+    public void prepareForUpdate(Context context, int appWidgetId, SuntimesMoonData data)
     {
         // EMPTY
     }