+++ app/src/main/java/com/forrestguice/suntimeswidget/settings/WidgetSettings.java
@@ -95,6 +95,9 @@ public class WidgetSettings
     public static final String PREF_KEY_APPEARANCE_TIMEFORMATMODE = "timeformatmode";
     public static final TimeFormatMode PREF_DEF_APPEARANCE_TIMEFORMATMODE = TimeFormatMode.MODE_SYSTEM;
 
+    public static final String PREF_KEY_GENERAL_RISESETORDER = "risesetorder";
+    public static final RiseSetOrder PREF_DEF_GENERAL_RISESETORDER = RiseSetOrder.TODAY;
+
     public static final String PREF_KEY_GENERAL_TIMEMODE = "timemode";
     public static final TimeMode PREF_DEF_GENERAL_TIMEMODE = TimeMode.OFFICIAL;
 
@@ -187,6 +190,9 @@ public class WidgetSettings
 
     public static final String PREF_KEY_DATE_DAY = "dateDay";
     public static final int PREF_DEF_DATE_DAY = -1;
+
+    public static final String PREF_KEY_NEXTUPDATE = "nextUpdate";
+    public static final long PREF_DEF_NEXTUPDATE = -1L;
 
     ///////////////////////////////////////////////////////////////////////////////////////////////
     ///////////////////////////////////////////////////////////////////////////////////////////////
@@ -939,6 +945,98 @@ public class WidgetSettings
             BLUE4.setDisplayStrings( context.getString(R.string.timeMode_blue4_short),
                     context.getString(R.string.timeMode_blue4) );
         }
+    }
+
+
+    /**
+     * RiseSetOrder
+     */
+    public static enum RiseSetOrder
+    {
+        TODAY("Today"),
+        LASTNEXT("Last / Next");
+
+        private String displayString;
+
+        private RiseSetOrder(String displayString)
+        {
+            this.displayString = displayString;
+        }
+
+        public String toString()
+        {
+            return displayString;
+        }
+
+        public void setDisplayString(String displayString)
+        {
+            this.displayString = displayString;
+        }
+
+        public static void initDisplayStrings( Context context )
+        {
+            TODAY.setDisplayString( context.getString(R.string.risesetorder_today) );
+            LASTNEXT.setDisplayString( context.getString(R.string.risesetorder_lastnext) );
+        }
+    }
+
+    ///////////////////////////////////////////////////////////////////////////////////////////////
+    ///////////////////////////////////////////////////////////////////////////////////////////////
+
+    public static long getNextSuggestedUpdate(Context context, int appWidgetId)
+    {
+        SharedPreferences prefs = context.getSharedPreferences(PREFS_WIDGET, 0);
+        String prefs_prefix = PREF_PREFIX_KEY + appWidgetId;
+        return prefs.getLong(prefs_prefix + PREF_KEY_NEXTUPDATE, -1);
+    }
+    public static void saveNextSuggestedUpdate(Context context, int appWidgetId, long updateTime)
+    {
+        SharedPreferences.Editor prefs = context.getSharedPreferences(PREFS_WIDGET, 0).edit();
+        String prefs_prefix = PREF_PREFIX_KEY + appWidgetId;
+        prefs.putLong(prefs_prefix + PREF_KEY_NEXTUPDATE, updateTime);
+        prefs.apply();
+    }
+    public static void deleteNextSuggestedUpdate(Context context, int appWidgetId)
+    {
+        SharedPreferences.Editor prefs = context.getSharedPreferences(PREFS_WIDGET, 0).edit();
+        String prefs_prefix = PREF_PREFIX_KEY + appWidgetId;
+        prefs.remove(prefs_prefix + PREF_KEY_NEXTUPDATE);
+        prefs.apply();
+    }
+
+    ///////////////////////////////////////////////////////////////////////////////////////////////
+    ///////////////////////////////////////////////////////////////////////////////////////////////
+
+    public static void saveRiseSetOrderPref(Context context, int appWidgetId, RiseSetOrder mode)
+    {
+        SharedPreferences.Editor prefs = context.getSharedPreferences(PREFS_WIDGET, 0).edit();
+        String prefs_prefix = PREF_PREFIX_KEY + appWidgetId + PREF_PREFIX_KEY_GENERAL;
+        prefs.putString(prefs_prefix + PREF_KEY_GENERAL_RISESETORDER, mode.name());
+        prefs.apply();
+    }
+    public static RiseSetOrder loadRiseSetOrderPref(Context context, int appWidgetId)
+    {
+        SharedPreferences prefs = context.getSharedPreferences(PREFS_WIDGET, 0);
+        String prefs_prefix = PREF_PREFIX_KEY + appWidgetId + PREF_PREFIX_KEY_GENERAL;
+        String modeString = prefs.getString(prefs_prefix + PREF_KEY_GENERAL_RISESETORDER, PREF_DEF_GENERAL_RISESETORDER.name());
+
+        RiseSetOrder mode;
+        try
+        {
+            mode = RiseSetOrder.valueOf(modeString);
+
+        } catch (IllegalArgumentException e) {
+            mode = PREF_DEF_GENERAL_RISESETORDER;
+            Log.w("loadRiseSetOrder", "Failed to load value '" + modeString + "'; using default '" + PREF_DEF_GENERAL_RISESETORDER.name() + "'.");
+        }
+        return mode;
+    }
+    public static void deleteRiseSetOrderPref(Context context, int appWidgetId)
+    {
+        SharedPreferences.Editor prefs = context.getSharedPreferences(PREFS_WIDGET, 0).edit();
+        String prefs_prefix = PREF_PREFIX_KEY + appWidgetId + PREF_PREFIX_KEY_GENERAL;
+        prefs.remove(prefs_prefix + PREF_KEY_GENERAL_RISESETORDER);
+        prefs.apply();
     }
 
     ///////////////////////////////////////////////////////////////////////////////////////////////
@@ -2153,6 +2251,7 @@ public class WidgetSettings
 
     public static void deletePrefs(Context context, int appWidgetId)
     {
+        deleteNextSuggestedUpdate(context, appWidgetId);
         deleteActionModePref(context, appWidgetId);
         deleteActionLaunchPref(context, appWidgetId);
 
@@ -2174,6 +2273,7 @@ public class WidgetSettings
         deleteTimeMode2Pref(context, appWidgetId);
         deleteTimeMode3Pref(context, appWidgetId);
 
+        deleteRiseSetOrderPref(context, appWidgetId);
         deleteCompareModePref(context, appWidgetId);
         deleteShowComparePref(context, appWidgetId);
         deleteShowNoonPref(context, appWidgetId);
@@ -2226,5 +2326,6 @@ public class WidgetSettings
         SolarTimeMode.initDisplayStrings(context);
         DateMode.initDisplayStrings(context);
         TimeFormatMode.initDisplayStrings(context);
+        RiseSetOrder.initDisplayStrings(context);
     }
 }